Fixing things

+++ src/lib/syntax/ast/pos.rs
@@ -1,14 +1,14 @@
-#[deriving(Clone, PartialEq)]
+#[derive(Clone, PartialEq)]
 /// A position in Javascript source code
 pub struct Position {
     /// The column number
-    pub column_number : uint,
+    pub column_number : u64,
     /// The line number
-    pub line_number : uint
+    pub line_number : u64
 }
 impl Position {
     /// Create a new position
-    pub fn new(line_number: uint, column_number: uint) -> Position {
+    pub fn new(line_number: u64, column_number: u64) -> Position {
         Position {
             line_number: line_number,
             column_number: column_number
